Refactor downloaders package
