+++ downloaders/ftp.go
@@ -1,4 +1,4 @@
-package core
+package downloaders
 
 import (
 	"net/url"
@@ -10,6 +10,7 @@ import (
 
 	"github.com/secsy/goftp"
 	"github.com/snickers/snickers/db"
+	"github.com/snickers/snickers/helpers"
 	"github.com/snickers/snickers/types"
 )
 
@@ -26,9 +27,9 @@ func FTPDownload(logger lager.Logger, dbInstance db.Storage, jobID string) error
 		return err
 	}
 
-	job.LocalSource = GetLocalSourcePath(job.ID) + path.Base(job.Source)
-	job.LocalDestination = GetLocalDestination(dbInstance, jobID)
-	job.Destination = GetOutputFilename(dbInstance, jobID)
+	job.LocalSource = helpers.GetLocalSourcePath(job.ID) + path.Base(job.Source)
+	job.LocalDestination = helpers.GetLocalDestination(dbInstance, jobID)
+	job.Destination = helpers.GetOutputFilename(dbInstance, jobID)
 	job.Status = types.JobDownloading
 	job.Details = "0%"
 	dbInstance.UpdateJob(job.ID, job)
